refactor(cli/amap): move discovery to rankings
