+++ tosho/src/impl/amap/mod.rs
@@ -9,6 +9,7 @@ pub(crate) mod config;
 pub(crate) mod download;
 pub(crate) mod manga;
 pub(crate) mod purchases;
+pub(crate) mod rankings;
 
 #[derive(Subcommand)]
 pub(crate) enum AMAPCommands {
